+++ index.html
@@ -9,6 +9,32 @@
   <body>
     <div class="todoapp">
       <h1>TODOS</h1>
+      <!-- <form id="todo-form">
+        <input
+          id="new-todo-title"
+          class="new-todo"
+          placeholder="할일을 추가해주세요"
+          autofocus
+        />
+      </form>
+      <main>
+        <input class="toggle-all" type="checkbox" />
+        <ul id="todo-list" class="todo-list"></ul>
+        <div class="count-container">
+          <span class="todo-count">총 <strong>0</strong> 개</span>
+          <ul class="filters">
+            <li>
+              <a class="all selected" href="#">전체보기</a>
+            </li>
+            <li>
+              <a class="active" href="#active">해야할 일</a>
+            </li>
+            <li>
+              <a class="completed" href="#completed">완료한 일</a>
+            </li>
+          </ul>
+        </div>
+      </main> -->
     </div>
     <script src="./index.js" type="module"></script>
   </body>
