+++ src/components/TodoList.js
@@ -1,4 +1,4 @@
-import { createElement } from "./utils/domCreator.js";
+import { createElement } from "../utils/domCreator.js";
 
 export default class TodoList {
   constructor({ $app, mode, list }) {
